Move MotionVector to mc.rs
